+++ packages/orq-rc/RELEASES.md
@@ -489,3 +489,13 @@ Based on:
 - [typescript v3.2.0-rc.7] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.7 - packages/orq-rc
+
+## 2025-02-19 07:00:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.495.1 (2.515.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.8 - packages/orq-rc